% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmEx.R
\docType{data}
\name{AmEx}
\alias{AmEx}
\title{American Express Credit Card data}
\format{
A data frame with 13444 observations and 56 variables
\describe{
  \item{DEFAULT}{observed outcome variable}
  \item{CARDHLDR}{selection indicator}
  \item{other variables}{use names(AmEx)}

}
}
\source{
<http://people.stern.nyu.edu/wgreene/Microeconometrics.htm>
}
\usage{
data(AmEx)
}
\description{
AmEx data consisted of 13,400 observations on credit card applications received
in a single month in 1988. Of the full sample, 10,499 applications were approved
and the next 12 months of spending and default behavior were observed.
Important variables in the data includes Demographic and socioeconomic factors of
the applicants (e.g. Age, Income, whether the applicant owns his or her home, whether
the applicant is self employed or not and  the number of dependents living with the applicant).
An important factor for granting the credit facility is grouped under ``Derogatories and Other Credit Data".
}
\examples{
data(AmEX)

\donttest{

outcome <- DEFAULT~AGE+ACADMOS+ADEPCNT+AEMPMOS+MAJORDRG+ MINORDRG+OWNRENT+
 APADMOS+AMAMIND+INCOME+SELFEMPL+ TRADACCT+ INCPER+ EXP_INC+CPTOPNB+
 CPTOPNG+ CPT30C+CPTF30+CPTAVRV+CBURDEN


selection <- CARDHLDR~AGE+ACADMOS+ADEPCNT+AEMPMOS+MAJORDRG+ MINORDRG+OWNRENT+
 APADMOS+AMAMIND+INCOME+SELFEMPL+ TRADACCT+ INCPER+CPTOPNB+
 CPTOPNG+ CPT30C+CPTF30+CPTAVRV+CBURDEN+BANKSAV+BANKCH+BANKBOTH+CREDMAJR+ACBINQ


pp <- HeckSelect(selection, outcome, data=AmEx, allowParallel = TRUE, penalty="ALASSO", Model="AMH",crit="bic")}

}
\references{
Greene WH (1998) Sample selection in credit-scoring models
}
\keyword{datasets}
