% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSelect.R
\name{boot_ProbitLasso}
\alias{boot_ProbitLasso}
\title{Title Bootstrap validation for probit regression with Lasso penalty}
\usage{
boot_ProbitLasso(object, data, mboot = 50, seed, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "ProbitLasso"}

\item{data}{data matrix containing both the outcome and selection variables}

\item{mboot}{number of bootstrap samples. 50 bootstrap samples are
used if not specified}

\item{seed}{integer value set for reproducibility}

\item{...}{}
}
\value{
coefficients: same as in coef

nonconvergence: the number of samples that did not converge
in the bootstrap sample

lambda: optimal shrinkage parameter

resu: data matrix containing apparent, bootstrap, test and optimism
corrected performance measures

Vectors of performance measures for each bootstrap sample are also returned
}
\description{
Title Bootstrap validation for probit regression with Lasso penalty
}
\examples{
pp2 <- ProbitLasso(formula, data=data, allowParallel = TRUE, penalty="ALASSO", crit="bic")
boot_ProbitLasso(pp2, data=data, mboot=2, seed=1)

}
