% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSelect.R
\name{bootValidate_Pval}
\alias{bootValidate_Pval}
\title{Title Bootstrap validation for object of model "HeckPval"(only implemented for Normal error distribution)}
\usage{
bootValidate_Pval(object, data, mboot = 50, seed, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "HeckPval"}

\item{data}{data matrix containing both the outcome and selection variables}

\item{mboot}{number of bootstrap samples. 50 bootstrap samples are
used if not specified}

\item{seed}{integer value set for reproducibility}

\item{...}{}
}
\value{
coefficients: same as in coef

nonconvergence: the number of samples that did not converge
in the bootstrap sample

alpha: significance level: validation is based on model fitting without variable selection when alpha=1

resu: data matrix containing apparent, bootstrap, test and optimism
corrected performance measures

Vectors of performance measures for each bootstrap sample are also returned
}
\description{
Title Bootstrap validation for object of model "HeckPval"(only implemented for Normal error distribution)
}
\examples{
res <- HeckPval(selection, outcome, data=data,alpha=0.05)
bootValidate_Pval(res, data=data, mboot=2, seed=1)

}
