% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSelect.R
\name{HeckPval}
\alias{HeckPval}
\title{Title Variable selection using P-value in binary sample selection model}
\usage{
HeckPval(selection, outcome, data = sys.frame(sys.parent()), alpha = 0.05, ...)
}
\arguments{
\item{selection}{selection equation}

\item{outcome}{outcome equation}

\item{data}{data matrix containing both the outcome and selection variables}

\item{alpha}{significance level: default is 0.05}

\item{...}{}
}
\value{
class HeckPval containing variables selected via P-value and
the parameters supplied for the creation of the object. Function call is
also returned
}
\description{
Title Variable selection using P-value in binary sample selection model
}
\examples{
res <- HeckPval(selection, outcome, data=data,alpha=0.05))

}
