% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckSelect.R
\name{logLik.HeckSelect}
\alias{logLik.HeckSelect}
\title{Log-likelihood of a HeckSelect fit}
\usage{
\method{logLik}{HeckSelect}(object, ...)
}
\arguments{
\item{object}{a HeckSelect fit.}

\item{...}{ignored.}
}
\value{
an object of class "logLik".
}
\description{
Log-likelihood of a HeckSelect fit
}
