% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeBPSS.R
\name{logLik.RidgeBPSS}
\alias{logLik.RidgeBPSS}
\title{Log-likelihood of a RidgeBPSS fit}
\usage{
\method{logLik}{RidgeBPSS}(object, ...)
}
\arguments{
\item{object}{a RidgeBPSS fit.}

\item{...}{ignored.}
}
\value{
an object of class "logLik".
}
\description{
Log-likelihood of a RidgeBPSS fit
}
