% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckSelect.R
\name{print.HeckSelect}
\alias{print.HeckSelect}
\title{Print a HeckSelect fit}
\usage{
\method{print}{HeckSelect}(x, ...)
}
\arguments{
\item{x}{a HeckSelect fit.}

\item{...}{ignored.}
}
\value{
invisibly returns the fit.
}
\description{
Print a HeckSelect fit
}
