% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeBPSS.R
\name{print.RidgeBPSS}
\alias{print.RidgeBPSS}
\title{Print a RidgeBPSS fit}
\usage{
\method{print}{RidgeBPSS}(x, ...)
}
\arguments{
\item{x}{a RidgeBPSS fit.}

\item{...}{ignored.}
}
\value{
invisibly returns the fit.
}
\description{
Print a RidgeBPSS fit
}
