% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSelect.R
\name{ProbitLasso}
\alias{ProbitLasso}
\title{Title Lasso and Adaptive Lasso Probit Regression}
\usage{
ProbitLasso(
  formula,
  data = sys.frame(sys.parent()),
  lambda = NULL,
  allowParallel = FALSE,
  penalty = c("LASSO", "ALASSO"),
  crit = c("bic", "aic", "gcv"),
  ...
)
}
\arguments{
\item{formula}{model formula for probit regression. This is the outcome model when the missing data is deleted}

\item{data}{data matrix containing the outcome and covariates for probit regression}

\item{lambda}{shrinkage parameter, both scalar and vector are acceptable
When lambda=NULL, the internal vector of Lambda is used}

\item{allowParallel}{If true, the "doParallel" package is invoked}

\item{penalty}{can be ALASSO (for adaptive lasso) or LASSO (for Lasso) penalty}

\item{crit}{can be BIC, AIC or GCV, default is BIC}

\item{...}{}
}
\value{
class ProbitLasso containing penalized probit coefficients
similar to the GLMNET package results. Linear predictor and fitted values are returned
}
\description{
Title Lasso and Adaptive Lasso Probit Regression
}
\examples{
ProbitLasso(formula, data=data, allowParallel = TRUE, penalty="ALASSO", crit="bic")

}
