% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckSelect.R
\name{summary.HeckSelect}
\alias{summary.HeckSelect}
\title{Summarise a HeckSelect fit}
\usage{
\method{summary}{HeckSelect}(object, ...)
}
\arguments{
\item{object}{a HeckSelect fit.}

\item{...}{ignored.}
}
\value{
invisibly returns the fit.
}
\description{
Summarise a HeckSelect fit
}
