% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeBPSS.R
\name{summary.RidgeBPSS}
\alias{summary.RidgeBPSS}
\title{Summarise a RidgeBPSS fit}
\usage{
\method{summary}{RidgeBPSS}(object, ...)
}
\arguments{
\item{object}{a RidgeBPSS fit.}

\item{...}{ignored.}
}
\value{
invisibly returns the fit.
}
\description{
Summarise a RidgeBPSS fit
}
