% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSelect.R
\name{coef.HeckSelect}
\alias{coef.HeckSelect}
\title{Title Coefficients from objects returned by HeckSelect function}
\usage{
\method{coef}{HeckSelect}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "HeckSelect"}

\item{...}{}
}
\value{
vector of coefficients for the selection and outcome models
}
\description{
Title Coefficients from objects returned by HeckSelect function
}
\examples{
coef(HeckSelect(selection, outcome, data=data, allowParallel = TRUE, penalty="ALASSO", Model="AMH",crit="bic"))
}
