% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSelect.R
\name{HeckSelect}
\alias{HeckSelect}
\title{Title Regularization method for binary Heckman selection model}
\usage{
HeckSelect(
  selection,
  outcome,
  data = sys.frame(sys.parent()),
  lambda = NULL,
  allowParallel = FALSE,
  penalty = c("LASSO", "ALASSO"),
  Model = c("Normal", "AMH"),
  crit = c("bic", "aic", "gcv"),
  ...
)
}
\arguments{
\item{selection}{selection equation}

\item{outcome}{outcome equation}

\item{data}{data matrix containing both the outcome and selection variables}

\item{lambda}{shrinkage parameter, both scalar and vector are acceptable
When lambda=NULL, the internal vector of Lambda is used}

\item{allowParallel}{If true, the "doParallel" package is invoked}

\item{penalty}{can be ALASSO (for adaptive lasso) or LASSO (for Lasso) penalty}

\item{Model}{can either be Normal error of AMH (Ali-Mikhail-Haq) copula function}

\item{crit}{can be BIC, AIC or GCV, default is BIC}

\item{...}{}
}
\value{
class HeckSelect containing penalized coefficients and
the parameters supplied for the creation of the object. Function call is
also returned
}
\description{
Title Regularization method for binary Heckman selection model
}
\examples{
HeckSelect(selection, outcome, data=data, allowParallel = TRUE, penalty="ALASSO", Model="AMH",crit="bic")

}
