% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSelect.R
\name{predict.HeckSelect}
\alias{predict.HeckSelect}
\title{Title Predict Method for HeckSelect Fits}
\usage{
\method{predict}{HeckSelect}(object, newdata = NULL, na.action = na.pass, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "HeckSelect"}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{na.action}{function determining what should be done with missing values in newdata. The default is to predict NA}

\item{...}{}
}
\value{
vector of predicted values for P(outcome=1|selection=1)
}
\description{
Title Predict Method for HeckSelect Fits
}
\examples{
pp <- HeckSelect(selection, outcome, data=data, allowParallel = TRUE, penalty="ALASSO", Model="AMH",crit="bic")
predict(pp)
}
